Implement proper channel allocation and dealloction
With this patch the muxer will allocate channels in a clean way, it will
not return a pts until the modem is really ready to operate on this
channel.  It will also retry the command several times until the channel is
allocated.

Channel deallocation is also greatly improved, now we send a correct code to
close a channel and will repeat several times if the packet somehow (e.g.
modem was in power-saving state) lost.