Add support for control and access over Unix Domain sockets.
Listen on /var/run/gsm-mux for connections.
Over these connections access commands
  reset_modem
  get_power
  set_power {0,1}
  alloc_channel
  set_name
  exit
  connect

"connect" provides the significant new functionality.
Further IO on that socket goes directly to a channel to
the GSM controller.  That way we can talk to it without needing
to use a pty - simply open as socket and use that.

EOF from a socket appears as a zero-byte read, so make sure to
check for that in pseudo_device_read.
Writing to a closed socket produces SIGPIPE, so we definitely don't want to
exit on that case.
In general, SIGPIPE is quite uninteresting, so universally ignore it.

Signed-off-by: Neil Brown <neilb@suse.de>